requirements.txt
discord.py[voice]
yt-dlp
PyNaCl
aiohttp
python-dotenv
flask
groq
Pillow